requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pyhd3eb1b0_1
aiohttp=3.7.3=py36h27cfd23_1
async-timeout=3.0.1=py36h06a4308_0
attrs=20.3.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py36h06a4308_0
brotlipy=0.7.0=py36h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.4.13=h06a4308_1
cachetools=4.2.1=pyhd3eb1b0_0
certifi=2020.12.5=py36h06a4308_0
cffi=1.14.5=py36h261ae71_0
chardet=3.0.4=py36h06a4308_1003
click=7.1.2=pyhd3eb1b0_0
coverage=5.4=py36h27cfd23_2
cryptography=2.9.2=py36h1ba5d50_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
cython=0.29.21=py36h2531618_0
dataclasses=0.8=pyh4f3eec9_6
decorator=4.4.2=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.27.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
grpcio=1.35.0=py36h2157cd5_1
idna=2.10=pyhd3eb1b0_0
idna_ssl=1.1.0=py36h06a4308_0
imageio=2.9.0=pypi_0
importlib-metadata=2.0.0=py_1
intel-openmp=2020.2=254
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
libidn2=2.3.0=h27cfd23_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
lpips=0.1.3=pypi_0
lz4-c=1.9.3=h2531618_0
markdown=3.3.3=py36h06a4308_0
matplotlib=3.3.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.2.1=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
multidict=4.7.6=py36h7b6447c_1
ncurses=6.2=he6710b0_1
nettle=3.7.2=hbbd107a_1
networkx=2.5=pypi_0
ninja=1.10.2=py36hff7bd54_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.1.0=py_0
olefile=0.46=py36_0
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h27cfd23_0
pillow=8.1.0=py36he98fc37_0
pip=21.0.1=py36h06a4308_0
protobuf=3.14.0=py36h2531618_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pyjwt=2.0.1=py36h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pypi_0
pysocks=1.7.1=py36h06a4308_0
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=pypi_0
pytorch=1.8.1=py3.6_cuda10.1_cudnn7.6.3_0
pywavelets=1.1.1=pypi_0
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7=pyhd3eb1b0_1
scikit-image=0.17.2=pypi_0
scipy=1.5.4=pypi_0
setuptools=52.0.0=py36h06a4308_0
six=1.15.0=py36h06a4308_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.3.0=pyh4dce500_0
tensorboard-plugin-wit=1.6.0=py_0
tifffile=2020.9.3=pypi_0
tk=8.6.10=hbc83047_0
torchvision=0.9.1=py36_cu101
tqdm=4.56.2=pypi_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.3=pyhd3eb1b0_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yarl=1.5.1=py36h7b6447c_0
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
